[ember] @ember/debug as source of types, ember as re-export
