+++ types/ember/tsconfig.json
@@ -20,6 +20,8 @@
             "@ember/engine/*": ["ember__engine/*"],
             "@ember/utils": ["ember__utils"],
             "@ember/utils/*": ["ember__utils/*"],
+            "@ember/debug": ["ember__debug"],
+            "@ember/debug/*": ["ember__debug/*"],
             "@ember/object": ["ember__object"],
             "@ember/object/*": ["ember__object/*"],
             "@ember/polyfills": ["ember__polyfills"]
