+++ types/ember__debug/container-debug-adapter.d.ts
@@ -1,2 +1,11 @@
-import Ember from 'ember';
-export default class ContainerDebugAdapter extends Ember.ContainerDebugAdapter { }
+import Resolver from '@ember/engine/-private/resolver';
+
+/**
+ * The ContainerDebugAdapter helps the container and resolver interface
+ * with tools that debug Ember such as the Ember Inspector for Chrome and Firefox.
+ */
+export default class ContainerDebugAdapter extends Object {
+    resolver: Resolver;
+    canCatalogEntriesByType(type: string): boolean;
+    catalogEntriesByType(type: string): string[];
+}
